requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
openai>=1.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.0
plotly>=5.17.0
kaleido>=0.2.1
requests>=2.31.0
pandas>=1.5.0,<2.1.0
webcolors>=1.13
scikit-learn>=1.3.0